Refactor socket message handling
