+++ server/controllers/REST/token.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import User from "../models/user";
+import User from "../../models/user";
 import jwt from "jsonwebtoken";
 
 export const generate = async (req: Request, res: Response): Promise<void> => {
