+++ server/controllers/REST/user.ts
@@ -1,4 +1,4 @@
-import User from "../models/user";
+import User from "../../models/user";
 import { Request, Response } from "express";
 
 export const create = async (req: Request, res: Response): Promise<void> => {
