+++ server/routes/user.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { create } from "../controllers/user";
+import { create } from "../controllers/REST/user";
 
 const router = express.Router();
